Remove non-restful route
Create a extra route for upload is unnecessary given the fact that
this web app is restful, use a http method that is sutable for upload
is enough.
P.S. The current method for uploading file is put, but this may violate
the restful policy of put method should be idempotent. Should change to
patch method in feature.